Add files of geometry solver
These three files include the ipynb notebook where the main solver locate. The .txt file is used as a bridge between mathematica and python. The mathematica files is used to plot the geometry of the crystal. In this version I just used the scipy package to numerically minimize the interaction potential.